+++ app-core/src/main/kotlin/co/anitrend/core/component/content/AniTrendContent.kt
@@ -18,16 +18,13 @@
 package co.anitrend.core.component.content
 
 import android.os.Bundle
-import android.view.View
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.viewbinding.ViewBinding
 import co.anitrend.arch.core.model.ISupportViewModelState
-import co.anitrend.arch.extension.ext.UNSAFE
 import co.anitrend.arch.extension.network.contract.ISupportConnectivity
 import co.anitrend.arch.extension.network.model.ConnectivityState
-import co.anitrend.arch.recycler.extensions.isEmpty
 import co.anitrend.arch.ui.fragment.SupportFragment
 import co.anitrend.core.android.binding.IBindingView
 import co.anitrend.core.android.koinOf
@@ -36,7 +33,7 @@ import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 import org.koin.androidx.scope.fragmentScope
-import org.koin.core.scope.KoinScopeComponent
+import org.koin.core.component.KoinScopeComponent
 import timber.log.Timber
 
 abstract class AniTrendContent<B : ViewBinding> : SupportFragment(),
@@ -44,7 +41,7 @@ abstract class AniTrendContent<B : ViewBinding> : SupportFragment(),
 
     override var binding: B? = null
 
-    override val scope by lazy(UNSAFE) { fragmentScope() }
+    override val scope by fragmentScope()
 
     /**
      * Additional initialization to be done in this method, this method will be called in
@@ -56,9 +53,8 @@ abstract class AniTrendContent<B : ViewBinding> : SupportFragment(),
      * @param savedInstanceState
      */
     override fun initializeComponents(savedInstanceState: Bundle?) {
-        lifecycleScope.launch {
+        lifecycleScope.launchWhenResumed {
             koinOf<ISupportConnectivity>().connectivityStateFlow
-                .flowWithLifecycle(lifecycle, Lifecycle.State.RESUMED)
                 .onEach { state ->
                     Timber.v("Connectivity state changed: $state")
                     if (state == ConnectivityState.Connected) viewModelState()?.retry()
