+++ app-core/src/main/kotlin/co/anitrend/core/component/sheet/AniTrendBottomSheet.kt
@@ -28,7 +28,6 @@ import androidx.annotation.MenuRes
 import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.viewbinding.ViewBinding
 import co.anitrend.arch.core.model.ISupportViewModelState
-import co.anitrend.arch.extension.ext.UNSAFE
 import co.anitrend.core.R
 import co.anitrend.core.android.binding.IBindingView
 import co.anitrend.core.android.components.sheet.SheetBehaviourCallback
@@ -37,14 +36,13 @@ import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.MainScope
 import org.koin.androidx.scope.fragmentScope
-import org.koin.core.scope.KoinScopeComponent
+import org.koin.core.component.KoinScopeComponent
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import android.widget.FrameLayout
 import androidx.activity.OnBackPressedCallback
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
-import co.anitrend.arch.extension.ext.getColorFromAttr
 import co.anitrend.arch.extension.network.contract.ISupportConnectivity
 import co.anitrend.arch.extension.network.model.ConnectivityState
 import co.anitrend.arch.ui.common.ILifecycleController
@@ -66,7 +64,7 @@ abstract class AniTrendBottomSheet<B : ViewBinding>(
 
     override var binding: B? = null
 
-    override val scope by lazy(UNSAFE) { fragmentScope() }
+    override val scope by fragmentScope()
 
     protected lateinit var behavior: BottomSheetBehavior<*>
 
@@ -108,9 +106,8 @@ abstract class AniTrendBottomSheet<B : ViewBinding>(
      * @param savedInstanceState
      */
     override fun initializeComponents(savedInstanceState: Bundle?) {
-        lifecycleScope.launch {
+        lifecycleScope.launchWhenResumed {
             koinOf<ISupportConnectivity>().connectivityStateFlow
-                .flowWithLifecycle(lifecycle, Lifecycle.State.RESUMED)
                 .onEach { state ->
                     Timber.v("Connectivity state changed: $state")
                     if (state == ConnectivityState.Connected) viewModelState()?.retry()
